+++ modules/Accounting/Expenses.php
@@ -143,13 +143,14 @@ if ( ! $_REQUEST['modfunc'] )
 		$i++;
 	}
 
-	if ( ! empty( $RET ) && ! $_REQUEST['print_statements'] && AllowEdit() )
+	$columns = [];
+
+	if ( ! empty( $RET )
+		&& ! $_REQUEST['print_statements']
+		&& AllowEdit()
+		&& ! isset( $_REQUEST['_ROSARIO_PDF'] ) )
 	{
-		$columns = [ 'REMOVE' => '' ];
-	}
-	else
-	{
-		$columns = [];
+		$columns = [ 'REMOVE' => '<span class="a11y-hidden">' . _( 'Delete' ) . '</span>' ];
 	}
 
 	$columns += [
@@ -158,8 +159,7 @@ if ( ! $_REQUEST['modfunc'] )
 		'COMMENTS' => _( 'Comment' ),
 	];
 
-	if ( empty( $_REQUEST['print_statements'] )
-		&& ! isset( $_REQUEST['_ROSARIO_PDF'] ) )
+	if ( empty( $_REQUEST['print_statements'] ) )
 	{
 		$columns += [ 'FILE_ATTACHED' => _( 'File Attached' ) ];
 	}