+++ modules/Accounting/StaffPayments.php
@@ -150,15 +150,14 @@ if ( UserStaffID() && ! $_REQUEST['modfunc'] )
 		$i++;
 	}
 
+	$columns = [];
+
 	if ( ! empty( $RET )
 		&& empty( $_REQUEST['print_statements'] )
-		&& AllowEdit() )
+		&& AllowEdit()
+		&& ! isset( $_REQUEST['_ROSARIO_PDF'] ) )
 	{
 		$columns = [ 'REMOVE' => '<span class="a11y-hidden">' . _( 'Delete' ) . '</span>' ];
-	}
-	else
-	{
-		$columns = [];
 	}
 
 	$columns += [
@@ -167,8 +166,7 @@ if ( UserStaffID() && ! $_REQUEST['modfunc'] )
 		'COMMENTS' => _( 'Comment' ),
 	];
 
-	if ( empty( $_REQUEST['print_statements'] )
-		&& ! isset( $_REQUEST['_ROSARIO_PDF'] ) )
+	if ( empty( $_REQUEST['print_statements'] ) )
 	{
 		$columns += [ 'FILE_ATTACHED' => _( 'File Attached' ) ];
 	}