+++ modules/Student_Billing/StudentPayments.php
@@ -201,7 +201,8 @@ if ( UserStudentID()
 
 	if ( ! empty( $RET )
 		&& empty( $_REQUEST['print_statements'] )
-		&& AllowEdit() )
+		&& AllowEdit()
+		&& ! isset( $_REQUEST['_ROSARIO_PDF'] ) )
 	{
 		$columns = [ 'REMOVE' => '<span class="a11y-hidden">' . _( 'Delete' ) . '</span>' ];
 	}
@@ -213,8 +214,7 @@ if ( UserStudentID()
 		'LUNCH_PAYMENT' => _( 'Lunch Payment' ),
 	];
 
-	if ( empty( $_REQUEST['print_statements'] )
-		&& ! isset( $_REQUEST['_ROSARIO_PDF'] ) )
+	if ( empty( $_REQUEST['print_statements'] ) )
 	{
 		$columns += [ 'FILE_ATTACHED' => _( 'File Attached' ) ];
 	}